JPA 조회(userInfo2.do) 수정 / 수정(changeUserInfo2.do) 추가

+++ src/main/java/com/eunjy/stockid/jpaDomain/user/UsrJpaVO.java
@@ -21,7 +21,7 @@ public class UsrJpaVO {
 	@Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "USR_NUM")
-    private Long usrNum;
+    private int usrNum;
 
     @Column(name = "USR_ID", length = 30, nullable = false)
     private String usrId;
